+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes of releases are documented in this file
 using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
+## [unreleased]
+
+### Added
+
+- Add new CLI command to generate media images.
+- Added new plugin configurations for default values.
+- Add new option to specify the sizes of generated product images.
+
 ## [1.0.0]
 
 ### Added